Refactor cookie handling in browser services to utilize background script for Chrome and Safari; streamline session loading logic
