+++ angular/src/app/services/browser.chrome.ts
@@ -1,21 +1,22 @@
 import browser from "webextension-polyfill";
+import { getCookiesFromBackground } from "./browser.background";
 
 export interface ChromeSessionLoaderContext {
-    getAllCookiesDirectly: () => Promise<browser.Cookies.Cookie[]>;
-    getCookiesFromBackground: () => Promise<browser.Cookies.Cookie[]>;
+    hostManage: string[];
+    hostApi: string[];
     processCookies: (cookies: browser.Cookies.Cookie[]) => void;
 }
 
 export function loadChromeSessions(context: ChromeSessionLoaderContext, cb: () => void): void {
-    context.getAllCookiesDirectly()
+    browser.cookies.getAll({})
         .catch(err => {
             console.warn("BrowserService.Chrome: direct getAll failed, falling back to SW:", err);
-            return context.getCookiesFromBackground();
+            return getCookiesFromBackground(context.hostManage, context.hostApi);
         })
         .then(cookies => {
             // getAll({}) resolves with [] (no reject) when host access isn't granted; fall back.
             if (!cookies || cookies.length === 0) {
-                return context.getCookiesFromBackground();
+                return getCookiesFromBackground(context.hostManage, context.hostApi);
             }
             return cookies;
         })
